LR5/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2
flask==2.3.3
flask-cors==4.0.0